engine-metrics-agent/src/ai/requirements.txt
google-cloud-aiplatform>=1.30.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
chess>=1.9.0
python-chess>=1.999
requests>=2.28.0
flask>=2.2.0
gunicorn>=20.1.0
python-dotenv>=0.19.0
google-cloud-storage>=2.5.0
google-cloud-firestore>=2.7.0
nltk>=3.7
matplotlib>=3.5.0
seaborn>=0.11.0